requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyJWT
python-dotenv
psycopg2-binary
azure-identity
pytz
bcrypt==3.2.0
ruff
